moved recipe components to dedicated folder
